Stop authorizing right after first login
Decidim is very usable without authorizations. We should not bother
users right after registration until they do something that requires
authorization. Otherwise we can let them enjoy all the decidim goodness
that don't require authorizations.

If apps wanted to override this, I think app they are still able to do
it by overriding the sessions controller, and they won't have to bother
with the number of enabled authorizations and so on like it happens when
we provide it in core.