Ship a thin (160Kb) instead of a fat (7.9Mb) gem
We were shipping the whole repository inside the decidim gem, but it was
due to a bug in the file regexp. However, that bug was actually keeping
generators (mainly) working since we used to depend on the internal
inter-gem directory structure.

This has been fixed now so we can go and fix the regexp bug.